nimbo-wms-168: Rename `NimboWMS.ServiceDefaults` to `Nimbo.Wms.ServiceDefaults` and update references
- Renamed project and namespace to align with naming conventions.
- Updated solution file and project references across all dependent projects.
