+++ Nimbo.Wms/Nimbo.Wms.csproj
@@ -20,7 +20,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\Nimbo.Wms.Infrastructure\Nimbo.Wms.Infrastructure.csproj" />
-      <ProjectReference Include="..\NimboWMS.ServiceDefaults\NimboWMS.ServiceDefaults.csproj" />
+      <ProjectReference Include="..\Nimbo.Wms.ServiceDefaults\Nimbo.Wms.ServiceDefaults.csproj" />
     </ItemGroup>
 
     <PropertyGroup>
